refactor: use hooks
